feat: added CTA, Projects, Services to home page
