+++ src/common/components/NextImage.tsx
@@ -1,0 +1,2 @@
+import Image from "next/image";
+export default Image;
