+++ src/widgets/Footer/components/Copyright.tsx
@@ -7,11 +7,11 @@ export default function Copyright() {
       <div className="row align-items-center">
         <div className="col-lg-6">
           <p className="fsz-16 cr-999">
-            Copyright: © 2025 by
+            Copyright: © 2025 by &nbsp;
             <Link href="/" className="text-white">
               Halocraft
             </Link>
-            All Rights Reserved.
+            &nbsp; All Rights Reserved.
           </p>
         </div>
       </div>
